requirements.txt
dashscope>=1.14.0
Pillow>=10.0.0
requests>=2.27.1
streamlit>=1.10.0
python-dotenv>=0.20.0
base64io>=1.0.3 